OpenCL stuff added

+++ src/kernel.cl
@@ -1,4 +1,4 @@
-__kernel void add(__global const float* a, __global const float* b, __global float* result) {
+__kernel void add(__global const int* a, __global const int* b, __global int* result) {
     int id = get_global_id(0);
     result[id] = a[id] + b[id];
 }
